DBus part of monitor configuration change
Added hooks to respond to the new power_manager::kSetScreenPowerSignal DBus signal.  This change doesn't start to act on this event but does expose how we will receive it in order to simplify a following change which does use it.

This change is expected to result in no behavioural change.

BUG=chromium:126493
TEST=Manually tested that brightness behaviour hasn't changed on a Lumpy.


Review URL: https://chromiumcodereview.appspot.com/10424009