ash: Change how mouse/touch-events are handled by the status tray popup and partial screenshot view.
The tray-bubble installs itself as a root-window event handler, intead of message-loop
observer, to close itself when clicked/tapped outside of the window. The partial
screenshot-view's event filter passes the event directly to the view, so that it
can process the event, and the event doesn't fall through to other filters.

BUG=120653
TEST=manually

Review URL: https://chromiumcodereview.appspot.com/9969045