compositor: minor fixes
This covers a few things, from failing to properly take into account damage in
certain cases (e.g. a layer with a texture but no delegate, or a texture layer
that is offset from its parent), or trying to set vsync on an offscreen context.

BUG=none
TEST=manual


Review URL: https://chromiumcodereview.appspot.com/10453016