Obey Wimax save credentials setting.
Note: the default for wimax->save_credentials() is false, so for a new wimax network the checkbox will be unchecked initially.
Also, it doesn't make any sense to "share" a network if credentials are not saved, so that will be hidden also when "save credentials" is hidden.

BUG=128429
TEST=Ensure that save credentials + share checkboxes are hidden during login and visible after login for secure Wimax networks.

Review URL: https://chromiumcodereview.appspot.com/10453010