Add NOTIFICATION_TAB_CONTENTS_DESTROYED when a TabContentsWrapper is about to be torn down. Code that wants to access TabContentsWrappers (and their members) safely should listen to this rather than NOTIFICATION_WEB_CONTENTS_DESTROYED.
This fixes a crash in GoogleURLTracker that tried to access the already-NULLed infobar_tab_helper() on the TCW. The other points changed in this CL probably work fine today, but using this notification is more future-proof if we muck with destruction order (and sometimes slightly simpler to boot). 

This also eliminates one use of NOTIFICATION_WEB_CONTENTS_DESTROYED entirely; there was some hacky code in task_manager_resource_providers.cc that tried to use it to backstop listening for another notification. Because the notification order is different with the new notification this wouldn't work; it also seems totally bogus anyway, so I just ripped it out. 

Finally, this adds a public getter for whether a TabContentsWrapper is being destroyed, and switches most places that were checking for the similar WebContents::IsBeingDestroyed() flag to use this instead.  This is usually more correct (since a number of places would check for WebContents destruction but if not found start accessing pieces of the TabContentsWrapper).

BUG=128512
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10445028