Remove the enable_metro GYP flag, effectively making it on by default.
The flag was mostly redundant with the enable_touch_ui flag, the only
additional effect of turning it on was to change the behavior of
ui::GetDisplayLayout.

If we later wanted to duplicate the effect that disabling the
enable_metro flag prior to this change would have, we can simply
change the condition in common.gypi that determines whether to turn on
the enable_touch_ui flag.

BUG=none


Review URL: https://chromiumcodereview.appspot.com/10440037