Add test for DeviceUserWhitelist and DeviceAllowNewUsers policies
This CL adds a PyAuto test for the closely interrelated
DeviceUserWhitelist and DeviceAllowNewUsers policies that control who
can log into a Chrome OS device.
As a prerequisite, the browser automation framework is fixed so that it
returns an error code when login is prevented by the whitelist rather
than timing out.

BUG=129722
TEST=New PyAuto ChromeosDevicePolicy.testUserWhitelistAndAllowNewUsers


Review URL: https://chromiumcodereview.appspot.com/10443028