Include Android-specific dependencies in the main DEPS file.
In order to be able to leverage versioning (in terms of Chromium rolls within
WebKit) and improve general maintainability, Android will move towards
deprecating the android.deps.git repository -which is a second gclient solution-
in favor of using gclient's target_os=[] functionality. This allows our
custom dependencies to be listed in the DEPS file as well.

No one is using this setting yet, but WebKit Chromium-Android checkouts will
be switched over shortly after this lands, which will be followed by the build
bots and developers working on Chrome on Android.

BUG=129577
TEST=

Review URL: https://chromiumcodereview.appspot.com/10445018