Use HTTPS when responding to user feedback.
This is not a particularly sensitive action, but there are (hopefully
low-sensitivity) cookies in the scope of the origin, and it's a generally
good idea to normalize the use of HTTPS over HTTP (as we have done with e.g.
"report a translation error". This issue was reported to us by Emanuel B
e3amn2l@gmail.com (no bug filed).

TEST=The response page still loads

Review URL: https://chromiumcodereview.appspot.com/10456014