Better cross platform behavior for blur.
On a focus event, we keep some state regarding the currently focused item in BrowserAccessibilityManager.

In doing so, we can sometimes cause some accessibility clients to jitter when focus moves from element to element (not via the tab order).

This patch removes the updating of the focused item within BrowserAccessibilityManager. Instead it prefers to simply fire the notification on the root for platforms that require it.


BUG=none
TEST=manually on Windows with NVDA.


Review URL: https://chromiumcodereview.appspot.com/10453070