libvpx Chrome build on Windows
This change allows libvpx to build on Windows with custom build step to
generate assembly offsets.

Added a tool with this change:

- copy_obj.sh

This tool copies a file a list of candidate locations. Such a tool is needed
because there are multiple possible locations where a .obj file is located,
depending on MSVS or Ninja is used.

Review URL: https://chromiumcodereview.appspot.com/10459033