When signing in to sync via the settings page, make sure to reset the member variables of SyncSetupHandler when the sign in overlay is closed, otherwise the user can get stuck in an error loop.
This problem appears only when trying OTP sign in.

BUG=129505
TEST=Try to signin using the signin overlay in the settings page.  Make a
few failed attempts until you see a captcha error.  Then close the overlay
and reopen it.  Make sure the captcha error is not shown and the user is
asked to enter username and password again.


Review URL: https://chromiumcodereview.appspot.com/10459030