Make ResourceHandler be non-refcounted.  Eliminate the OnRequestClosed method in favor of ~ResourceHandler.
This required making a couple ResourceHandlers support weak pointers.

The only non-trivial changes are related to DownloadResourceHandler.
The StartOnUIThread method becomes a static StartDownloadOnUIThread
method, and the call to SetDownloadID and CallStartedCB are now hidden
behind a Callback<void(DownloadId)>, which gets posted to the IO thread.

R=rdsmith@chromium.org,jam@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10455009