ash: Fix painting for windows with composited-views during hide animation.
The layers for the aura::Window and its child windows are acquired when the
window is hidden so that it can paint correctly during the hide animation.
However, if the window had views with layers, then those layers are not
retained, so the painting for these views do not happen correctly. To fix this,
take ownership of these layers as well when a window is hidden.

BUG=130219
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10444110