Change the visual treatment of panels titlebar on OSX per new UI design.
This introduces the behavior when closing the last panel will bring the existing Chrome window to the top. This happens because the 'close' button used is not the standard OSX close button which OSX knows in the face. The workaround to prevent this is coming in a separate patch.

Review URL: https://chromiumcodereview.appspot.com/10449104