Show sync promo even when restoring URLs
Currently we don't show the sync promo if the "Open a specific page or set of pages" preference is set.

Unfortunately this preference is used by distribution builds. This means that the sync promo never shows up in distribution builds.

This CL changes the startup logic so that the sync promo can be shown even if URLs are being restored on startup.

BUG=130641
TEST=


Review URL: https://chromiumcodereview.appspot.com/10442130