ash: Fix closing uber-tray popup by clicking on the tray.
If the user clicks on the tray while the bubble is up, then the bubble will
close. But after the mouse-up event on the tray, the bubble will show up
again. To prevent this from happening, if this mouse-press event that
closed the bubble is on the system tray, then do not let the event reach
the tray.

BUG=130691
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10440132