Fix page_cycler_dep setup_test_links.sh to be more robust
In various cases, including the one where the page_cycler data isn't
present, the setup_test_links.sh fails because there's a dangling
symlink which is not pointing to a directory.  This fixes it by
changing the test to just look for the existence of that file,
regardless of it's ulimate type.

BUG=chromium-os:31310
TEST=run desktopui_PyAutoPerfTests built from a external manifest twice

R=nirnimesh@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10465002