Teaching BrowsingDataRemover how to delete application data.
Adding a parameter to BrowsingDataRemover::Remove in order to support a
"Clear app data" flag in the "Clear Browsing Data UI". Additionally, this CL pulls the implementation of DomStorageContext::DeleteDataModifiedSince out of DomStorageContext, and into BrowsingDataRemover in order to manage the complexities of decision-making in the BDR, rather than exporting them out of chrome and into webkit.

This patch only addresses localStorage/sessionStorage and quota managed data. Cookies should already ignore the special storage policy, but we'll need to examine the remainder of the types that the BDR touches.

BUG=116372
TEST=unit_tests,test_shell_tests

Review URL: https://chromiumcodereview.appspot.com/10413072