Stop dropping buffering updates on the floor for small media resources.
Before, BufferedDataSource would fail to report buffering bytes earlier than the
current seek target.  This is wrong because BufferedResourceLoader has a
kForwardWaitThreshold buffer and if the seek is within that range, it'll just
keep feeding bytes to BDS until the seek target is reached.  Account for this.

BUG=130843


Review URL: https://chromiumcodereview.appspot.com/10503004