Fix about:tracing artefacts when scrolling around
Due to a typo in the tracing controller, we would take traces
that have no system traces and put them both into the regular trace slot
and the systemTrace slot. We would then import the regular trace twice. Amusingly,
this was mostly invisible to the rest of the system. The FastRectRenderer, however,
makes an assumption that the rectangles it receives always go left-to-right. When that
assumption is broken, it simply draws a big rect. So, when the zoom level was exactly right,
we would get a pair of slices that started on the right, then moved to the left. They would get fused
and a giant rect would cover the screen.

This is a quick fix. crbug.com/131145 is for preventative measures.

BUG=130639


Review URL: https://chromiumcodereview.appspot.com/10533002