Fix client-side phishing detection test flakiness and ChromeOS failure.
The problem is that RenderViewFakeResourcesTest creates a RenderThreadImpl, but without setting up the sandbox environment, specifically the domain socket used on Linux.  This caused the sandbox's pre-assigned file descriptor to be taken by other files or sockets, leading to a hang when the sandbox code tries to talk to the browser.

To fix this, add a way for tests to disable the WebSandboxSupport implementation in RendererWebKitPlatformSupportImpl.

The problem also affects WebRTCAudioDeviceTest, so fixed it there as well.

BUG=128705


Review URL: https://chromiumcodereview.appspot.com/10449094