Adding metric to tracking the lock screen request code path
For locking the screen we enter a code path were we do a method call, get a
empty reply, and then receive a signal to indicate that chrome can perform the
UI actions for screen locking. We have encountered issues where communication on
this path has failed. A fix for the specific bug has been landed, but we would
still like to track information about how we progress through this path to
detect future issues.

The metric that is being introduced is
"LockScreen.LockScreenPath". This metric replaces the existing
"LockScreen.RequestLockScreen" metric.

BUG=chromium-os:31346
TEST=Built Chrome with CL.
     Built ChromeOS image using CL.
     Ran image to confirm that no regressions had been introduced.
     Ran image and locked screen. Checked histogram populated.


Review URL: https://chromiumcodereview.appspot.com/10543004