Add a CachingValueStore class, which acts as a frontend for LeveldbValueStore, for use as an extension value store.
This class does not implement the ValueStore interface, because I want to avoid
the extra copying of Values that interface requires. This is a simple key/value
store.

I also modified LeveldbValueStore and its interface a bit to make optional some of the extra copying done on Write operations.

BUG=123366
TEST=no
TBR=akalin

Review URL: https://chromiumcodereview.appspot.com/10512024