Refactor painting of NetLogs.
The goal of this refactor is to
  (a) make it easier to add new custom parameter printers
  (b) be more resilient to bad inputs.

The visual look is unchanged, this is just refactoring the implementation. The only visual differences should be when visualizing unexpected (i.e. malformed) inputs. Before those parameters would be hidden, whereas now they are printed with default formatting.


Review URL: https://chromiumcodereview.appspot.com/10536023