ash: Adjust maximize/restore animation timing based on window size
In particular, make the animation faster if the window almost fills the screen and is being maximized.  This is a common case for freshly created windows.  At the default 400 ms timing this feels too slow, so this patch scales the time between 200 and 400 ms.

BUG=131016
TEST=added


Review URL: https://chromiumcodereview.appspot.com/10537029