[Mac] Log stack trace for OnChannelError() in bug 97285.
The hypothesis in comment #67 is that OnChannelError() is being called
to invalidate a channel, and the invalid channel is later being vended
from the cache.  This will verify if OnChannelError() was previously
called on the channel, and also provide a backtrace to determine what
the context was.

BUG=97285
TEST=crash server


Review URL: https://chromiumcodereview.appspot.com/10544038