gdata: Clean up "Uninteresting mock function call" from GDataFileSystemTest
After crrev.com/139288, the cache initialization is always initiated in
GDataFileSystem::Initialize(). We should ensure that the cache initialization
is complete in every test. Otherwise, completion of the cache initialization
is nondeterministic, hence "Uninteresting mock function call".

OnDirectoryChanged() in GDataFileSystemTest.CreateDirectoryWithService is
more interesting. We should set the expectation for this, as the contents
of the root directory is actually changed.

To prevent the pollution by "Uinteresting mock function call", wrap the
existing mock objects with StrictMock<> so that Uninteresting mock function
calls produce errors.


BUG=131494
TEST=out/Release/unit_tests --gtest_filter=GDataFileSystemTest.*


Review URL: https://chromiumcodereview.appspot.com/10534029