Merge trunk r140955 to the 20.0.1132 branch.
ninja and maybe make had made symbolic links when this was formerly a 'copies'
section, and you can't use cp to copy a file over (a symbolic link to) itself.
Dump the entire directory and rebuild it from scratch when
copy_keychain_reauthorize runs. This is not needed for the Xcode build, so it
should not need to be merged to any branches.

BUG=116210
TEST=none
Review URL: https://chromiumcodereview.appspot.com/10534035
Review URL: https://chromiumcodereview.appspot.com/10543055