Omnibox HistoryQuickProvider: add about:flag to control new scoring
Added an three-value about:flag to control whether the HistoryQuickProvider uses the new scoring function or the old.  The three values are new, old, and automatic.  For now automatic means old but in a later change I'll make automatic assign the status according to a field trial.

BUG=
TEST=by hand


Review URL: https://chromiumcodereview.appspot.com/10532051