Move the reboot hook out of remote_shutdown().
Alter the remote_shutdown hook to ensure we stop the
slave's reactor and note we need to reboot after
stopped.  Once the run() call returns, we call Reboot()
which actually handles triggering the reboot.

The purpose of this change is to avoid a possible race
that is causing the /sbin/shutdown command to sometimes
not run properly on Mac systems.  My theory (which
hasn't been verified) is that in all cases on Win and
Linux, /sbin/shutdown is getting called properly, but
in some cases on Mac, there's a race that can cause
/sbin/shutdown to not run properly.

The race theory is that we would call reactor.stop()
and then from the caller which invoked reactor.stop()
we would call /sbin/shutdown.  Since the caller is
running in a Twisted service, calling reactor.stop()
may sometimes be fast enough to stop even the service
that was about to call /sbin/shutdown.

By moving /sbin/shutdown out of the Twisted service
and running it only after the Buildbot process in
run_slave.py has come to a stop (ie. twistd.run()
has returned), we eliminate the possibility that a
race between the reactor.stop() and /sbin/shutdown
calls is the cause of the Mac failures to reboot
properly.

I change the subprocess.call()s to Popen().  When call()
was run, then Reboot() would wait for the /sbin/shutdown
call to return.  Meanwhile /sbin/shutdown will kill
this process and generate the interrupt exception in the
log.  Since waiting for /sbin/shutdown to return is not
important, use Popen() instead to remove the spurious
Exception from the log.

Add an explanation for the remaining common exception
message that appears in the twistd.log file when
remote_shutdown() is called by new masters on old slaves.

R=maruel@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10510008