mac: Make the backing store hidpi aware.
The calling code always passes "1" as scale factor, so this has
no observable effect for now.

The general convention is that everything that's IPCd to the
renderer is in view units, and the backing store and transport
dibs are the only things that use pixels.

BUG=31960
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10546058