Fix chrome-bot rev skip code.
The previous version looked for 'author' as
an attribute on the logentry XML tag (like
how 'revision' is an attr there).  This is
wrong.  Instead, the author appears as the
content of an enclosed <author></author> tag.

Use the _get_text() helper method to return
such a tag's contents and compare against that,
instead.

R=nsylvain@chromium.org
BUG=107678


Review URL: https://chromiumcodereview.appspot.com/10542067