Update touch_200_percent assets
These new assets are from the zip file attached to issue 131658.

Four of those images were renamed for consistency:
  theme_default_active.png => theme_frame_default_active.png
  theme_default_inactive.png => theme_frame_default_inactive.png
  incognito-switch-off.png => incognito_switch_off.png
  incognito-switch-on.png => incognito_switch_on.png

Five of them are new:
  otr_icon.png
  otr_icon_fullscreen.png
  incognito_background.png
  incognito_switch_off.png
  incognito_switch_on.png

In a separate CL I will update the GRD file to point to the first of these (currently it points to the version in touch_100_percent). The other four do not have resource IDs yet and the GRD file will not be updated yet.

TBR=ben@chromium.org
BUG=131658
TEST=Compare old and new images and check that they all look similar to each other

Review URL: https://chromiumcodereview.appspot.com/10546063