Detect output state at startup to avoid changes
Resetting the state of the outputs causes a black flash on the screen so avoid
doing this in the cases where the outputs were already configured such that they
were in a state we can recognize.

BUG=chromium:130943
TEST=Manually tested on lumpy with extra instrumentation to ensure that logging
in as guest does not change the CRTCs.


Review URL: https://chromiumcodereview.appspot.com/10532053