Disable the seccomp filter GPU process sandbox by default on Chrome OS.
We don't want the GPU process sandbox on by default on Chrome OS.
Since we control Chrome's command line, we can make sure it's enabled
for users, but disabled for developers.

Add an '--enable-chromeos-gpu-sandbox' flag, and remove the #ifdef that
disabled the sandbox on Chrome OS. Until we manually add the flag to
Chrome OS, the sandbox will continue to be disabled.

Moreover, even with the flag on, we only enable the sandbox on 64 bits and
for Intel GPUs.

BUG=127664
TEST=unit tests, and WebGL test websites on Lumpy.


Review URL: https://chromiumcodereview.appspot.com/10534049