Add relaunch prompt to linux help page.
This doesn't allow you to check for updates or show progress, it just notifies you if a new version is ready. As such, I didn't see a reason to limit it to Google Chrome builds.

This also fixes a bug where a 'disabled' state permanently hides the status text.

BUG=117403
TEST=manual with --check-for-update-interval=1 and manually editing the version


Review URL: https://chromiumcodereview.appspot.com/10541013