Refactor the content interface for RequestMediaAccessPermission.
This makes RequestMediaAccessPermission a part of WebContentsDelegate instead
of ContentBrowserClient. The request UI (or lack thereof) needs to be
controlled by the WebContentsDelegate so that platform apps can grant permission
differently than a normal browser. This also seems appropriate because
each request is made on behalf of a particular web contents.

BUG=130216
TEST=navigator.webkitGetUserMedia still works for normal web pages

Review URL: https://chromiumcodereview.appspot.com/10542092