Don't replace $1 with an empty string when localizing string for host plugin.
Previously chrome.l18n.getString() wouldn't touch $1 within the resource
when substitutions are not specified. This was fixed in crrev.com/141280, and
now it replaced $1 with an empty strings, which broke chromoting host
localization that relied on the previous incorrect behaviour. Fixed
the webapp to correctly handle strings substitutions.

BUG=132146

Review URL: https://chromiumcodereview.appspot.com/10533101