Let the layer animator's slow animation mode affect all animations.
Currently, this setting will only affect implicit animations. If you explicitly create an animation with a duration, it will not slow down. With this patch, we alter the duration of layer animation elements in the same way we do when disabling animations for tests: we hijack and possibly replace the duration in the layer animation element's constructor.

BUG=None
TEST=Manual -- hold shift when minimizing a window in aura.


Review URL: https://chromiumcodereview.appspot.com/10537143