Revert 141990 - Fix PPLT metrics in the RecentlyNavigated case.
When creating a dummy PrerenderContents, it's important that we not
fail for reasons like RecentlyNavigated, because this can cause us to
report incorrect statistics. This fix makes initialization of
PrerenderContents not fail for the dummy contents (which are never
started anyway).

As well, I added some unit tests around all of this functionality,
which were missing.

BUG=None


Review URL: https://chromiumcodereview.appspot.com/10544137

TBR=dominich@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10536154