Shutdown It2Me host UI only after the host is stopped.
Currently user interface object is destroyed before host is stopped.
The destructor for UserInterface object doesn't unregister that object
from list of host status observers, and it cannot do that because the 
destructor is executed on the UI thread. In the same time it's no 
longer necessary to destroy UI before the host is stopped. Changed
plugin destruction sequence so that UI is destroyed only after the 
host is stopped.

BUG=130596


Review URL: https://chromiumcodereview.appspot.com/10533138