mac: Make dynamic DPI changes work.
This fixes the software case completely, and makes everything
look good on the hardware path after a few frames. In the
hardware path, the IOSurfaces aren't tagged with their own DPI,
so they are rendered at the wrong scale for a few frames, until
the "DPI changed" message made it from the browser to the
renderer.

This CL adds three things that happen in response to DPI changes:
1.) It reallocates the browser-side BackingStoreMac at the new DPI.
2.) It lets the renderer do a full repaint after DPI changed.
3.) It passes the new DPI on to WebKit (in render_view_impl.cc)

(3) won't have an effect for hardware pages until
https://bugs.webkit.org/show_bug.cgi?id=88916 lands.

BUG=128267
TEST=change from lodpi and hidpi and back, check that software- and
hardware-rendered pages look good after the change in both foreground
and background tabs.

Review URL: https://chromiumcodereview.appspot.com/10548026