Add per-script font support for all ISO 15924 scripts.
Special script codes Zinh, Zxxx, Zzzz, Qaaa-Qabx are excluded.  Also, the
various Japanese script codes Hira, Kana, and Jpan are excluded in favor of the
single code Hrkt; likewise, for Korean, Kore is excluded in favor of Hang.
Hrkt and Hang are used just because they currently are already included; this
is planned to change in a separate patch
<https://chromiumcodereview.appspot.com/10532105>

BUG=2685
TEST=

Review URL: https://chromiumcodereview.appspot.com/10548021