Ensure that deletion of the default search engine does not disrupt sync.
Previously, it was possible for the deletion of the current default search
engine to cause the service to lose the pending sync default search engine.
Now, we only update the synced default search engine preference if we're not
already waiting for a sync preference.

The drawback here is that if for some reason sync gets into a situation where
we're stuck waiting for the new default search engine, the local default search
engine will not propagate to other clients. This is a situation where sync is
broken though, so sync problems are to be expected.

BUG=132446
TEST=unit_tests; manually adding a new default search engine and deleting the
old default while another client is online, then bringing the other client online.
The default should be preserved.

Review URL: https://chromiumcodereview.appspot.com/10544155