Do not re-use NaClBrokerHost objects if the broker is terminating.
When the last sel loader process dies, the nacl broker service sends
an IPC terminating the broker process. However the broker host object
which tracks that process stays around for a while until it is cleaned
up by the browser process, so subsequent launches attempt to use it to
launch new loaders, resulting in IPC pipe errors on win64.

This CL marks the broker host object as being in the process of
termination and does not re-use it.

R=mseaborn@chromium.org
BUG=none
TEST=nacl_integration

Review URL: https://chromiumcodereview.appspot.com/10443040