Protect against sync entries left by old clients
Old clients would sometimes end up with items that were deleted before
their existence was ever reported to the server.  For a while, we let
these entries accumulate in the database, and just filtered them out of
the list of items to commit when we created that list in GetCommitIds.

Then we learned that a subset of these entries could be deleted in such
a manner that they would be detected as database corruption.  The client
was udpated so it would delete these entries in a timely manner.  The
filter in GetCommitIds was removed.  See crbug.com/125381 for details.

It turns out that a lot of clients have some deleted local items in
their sync database created by older clients.  When they update to the
new client which doesn't have the GetCommitIds filter, they will try
to commit the local tombstones.  This is bad.

This commit addresses the problem in two ways:
1. Drop any server-unknown deleted items when we open the database.
2. As a fallback, put the filter back into GetCommitIds.  The filter
should not be required, so it will trigger a NOTREACHED() if it is
executed.

BUG=132905,125381
TEST=SyncableDirectoryTest.OldClientLeftUnsyncedDeletedLocalItem


Review URL: https://chromiumcodereview.appspot.com/10559018