Fix about:tracing
This changes our browser gyp to use the exact same build process as before
but just pull in about:tracing files from third_party/trace-viewer/src instead of
chrome/browser/resources/. This is a quick fix to restore about:tracing, which
as it turns out, people get upset about if it is broken. We can look for a more
"proper" fix with this landed.

BUG=132875


Review URL: https://chromiumcodereview.appspot.com/10566011