Clarify a warning message. Sort of.
It isn't possible to give an exact date for withdrawing
support because it is being phased in several steps.

But I agree with the general sentiment of the bug report, so
my plan is to write a blog post explaining the steps and
link to that.

BUG=132633
TBR=levin@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10555015