mac: Fix close button appearance in HiDPI.
Revert the explicit "scale" parts of r142304. In 10.7.3, the
backing scale factor did not apply to shadowOffset and
shadowBlurRadius. This appears fixed in 10.7.4, and retina
MBPs ship with 10.7.4 -- so the manual scaling caused the
scale to be applied twice, which moved the shadow off the
close button.

BUG=133134
TEST=look at a tab close button
TBR=rsesek

Review URL: https://chromiumcodereview.appspot.com/10556030