2x profile avatar icons
These are the profile avatar icons scaled to 2x. This will be used for HiDPI mode on Mac.

The _mac versions is the horizontally flipped version of the icon. I created it by doing the following:
  mogrify -flop *.png
I also ran pngcrush on all the icons.

This CL only contains the icons. The GRD changes are in this CL: https://chromiumcodereview.appspot.com/10559029/

BUG=132910
TEST=

Review URL: https://chromiumcodereview.appspot.com/10532189