Support for initializing filesystem when default download is drive.
When the default download location is a drive folder, we may be required to download a file to drive before the file manager has a chance to authenticate and initialize the fileystem. 
Reset gdata external data in DownloadItem for non-gdata downloads if the user switches to a local path.

BUG=133005,132810
TEST=manual.

Review URL: https://chromiumcodereview.appspot.com/10555013