Add a formal list of chrome-specific directories that cannot be tracked for now.
This is a side-effect that we set up so that the compiler slaves do not do a checkout of the test data for now. This may change eventually but it's not needed for now so just putting them in the untracked list is fine.

R=cmp@chromium.org
NOTRY=true
BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10556046