Upstream most of the user agent code for Android.
Brought along most of the command-line start-up switches.

As for requiring the InitUserAgent (and not just having the
logic in webkit/glue/user_agent), that's because currently unit tests
can't access Java code directly. Using pseudo dependency-injection
allows this to work for prod and test. Once apk-based tests are online,
we can remove this.


Review URL: https://chromiumcodereview.appspot.com/10473002