Make ResourceHandlers that defer loading also be responsible for calling Resume on their ResourceController.
What this does is avoid out-of-band calls to Resume the ResourceLoader, and
that means that an intermediate ResourceHandler (for example the
BufferedResourceHandler) will be able to implement ResourceController.

Consider the problematic scenerio that happens today:

1-  The BufferedResourceHandler consumes response headers and buffers network
data, delaying notifications to its downstream ResourceHandler.
2-  The BufferedResourceHandler decides if it is dealing with a download or
not, and optionally installs a DownloadResourceHandler as the downstream
handler.
3-  The BufferedResourceHandler replays the OnResponseStarted and
OnReadCompleted events to push data to its downstream handler.
4-  The downstream handler might decide to defer processing.
5-  When the downstream handler wants to resume processing, it needs to
tell someone to Resume.

At step 5, the complicated thing we do to ourselves today is we make that
Resume step tell the ResourceLoader to resume instead of the
BufferedResourceHandler.

This causes problems for the BufferedResourceHandler as it may start receiving
more network events before it has finished flushing its existing buffers to its
downstream handler.

This patch is a step toward fixing the above problem.  It makes all handlers
call Resume on their own controller when it is time to resume.  A subsequent
patch will leverage this to cleanup and simplify the BufferedResourceHandler
as well as the overall pause/resume semantics of the ResourceLoader.

Other cleanup in this patch:  The {Async,Sync,CrossSite}ResourceHandler
classes are now all constructed with an URLRequest.  This avoids messy
code to call ResourceDispatcherHostImpl::GetURLRequest().  Since the
lifetime of ResourceHandlers is now bound to the lifetime of their
associated URLRequest, this is a safe and simple change to make.  Other
ResourceHandlers, like the BufferedResourceHandler, were already holding
onto their associated URLRequest.

R=jam@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10568005