sync: don't inherit SupportsWeakPtr from ProfileSyncService
AutofillManager was the last use of this. This patch makes AutofillManager look up the sync service
directly instead of caching a pointer to it. Note that AutofillManager is ref counted and can outlive ProfileSyncService.

Since AutofillManager now asks ProfileSyncServiceFactory for a ProfileSyncService reference per-use, it may
return null if the profile is in incognito. UpdatePassphraseGenerationState didn't run into this before and
a null profile sync service meant something different -- updated that code accordingly.

I also made AutofillManager listen for TAB_CONTENTS_DESTROYED. I'm not sure if this was a bug in the wild (although A.M is ref counted, so maybe), but the TestTabContents infrastructure creates a scenario during SetUp where multiple AutofillManagers get created, and creating the second instance resets the TabContents that was used for the first instance. This makes certain methods on A.M instance 1's tab_contents_ pointer dangerous to call from ~AutofillManager, because parts of the owning TabContents have been destroyed by the time we get to ~AutofillManager.

The remaining use of WeakPtr was the FrontendDataTypeController, and I removed that use in http://src.chromium.org/viewvc/chrome?view=rev&revision=142836

BUG=100829
TEST=AutofillManagerTest

Review URL: https://chromiumcodereview.appspot.com/10584017