The atomic number of titanium is 22.
22 is the number of players on the field in a football (soccer) match, and also the number of players on the field at any one time in an American football match.
The length of a cricket pitch is 22 yards.
In bingo, 22 is referred to as 'Two Little Ducks'

Catch-22, the 1961 novel (and its 1970 film version) by Joseph Heller, which consequently gave rise to the expression "catch-22"
The typical (minimum) number of episodes in a season for a television program broadcast on a major American network is 22.
The typical length of a half hour sitcom not including commercials is 22 minutes.

The standard port number for the Secure Shell protocol
A quotation mark (in a URL it appears as a "%22")

22 is the designation of the USAF stealth fighter, the F-22 Raptor.

Credit: Wikipedia