Fix bug in EventFilter that arises when matchers are added to multiple events.
This change fixes a bug in EventFilter to handle the case where the existence of matching EventMatchers for event X causes matching against an event with name Y to crash. It also changes the representation of EventMatcher so that it owns the entire filter dictionary.


BUG=121479

Review URL: https://chromiumcodereview.appspot.com/10534125