Extensions Docs Server: DictGenerator
Implemented the HandlebarDictGenerator that works with the JSON Schema Compiler to
produce a dictionary that Handlebar can work with.

The HandlebarDictGenerator removes all items marked "nodoc", and only keeps the
necessary parts of the JSON. $refs can be resolved because, if the
dependencies are in other files, they are formatted as "dependency.Item"
(eg. windows.Window), so the link can be created based on this name.

APIDataSource now uses a DictGenerator instead of the raw JSON APIs.

BUG=131095

Review URL: https://chromiumcodereview.appspot.com/10577022