[Android] Send top-level data:// requests to the browser.
The Android WebView API informs the embedder of all top-level navigations,
including navigations to URLs with the data:// scheme.
To remain compatible with the WebView API (and the Android Browser
which is implemented using the WebView API) we need to be able to inform the
embedder of top-level navigations to data:// URLs too.
In order to do that we first need to send them over from the renderer to
the browser.

BUG=130006
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10440068