Re-factor TraySms to use ash::SmsObserver instead of NetoworkSmsHandler.
It turns out that the new SmsWatcher class does not work correctly if devices are added after it is initialized. This could be a problem with newly activated networks.

For now, revert to using the existing, tested SmsObserver class in Chrome through the existing ash::SystemTrayDelegate.

BUG=133569
TEST=See issue for testing details

Review URL: https://chromiumcodereview.appspot.com/10582028