Rely on losing focus, not on becoming unfullscreened, to destroy a fullscreen window.
On metacity we get spurious unfullscreen events that were preventing the window from going fullscreen, and alt-tab didn't cause the window to unfullscreen, so it wasn't destroyed by alt-tab. We should ignore the unfullscreen event and instead rely on FocusOut to determine when to destroy the window.

BUG=117021
TEST=


Review URL: https://chromiumcodereview.appspot.com/10542134