Use nv contrl x extension to query nvidia driver version.
So we can do it in browser process, not wait until gpu process launches and a gl context is created.

This is tested on my linux box with NVIDIA driver.

We need to put more code to tell chrome (Linux/NVIDIA) that we have enough information for blacklisting, so at GPU launch time we don't need to collect GPUInfo.

Will do it in a followup CL.

BUG=122912
TEST=tree green
TBR=jam
Review URL: https://chromiumcodereview.appspot.com/10536232