Full-screen Magnifier: Support warping the cursor on the edge
With this CL, a magnified region will moved when the cursor on the edge.

Only on debug build, we can test the this magnifier with "Ctrl + {" and "Ctrl + }" shortcut. These shortcut keys are temporary. These will be removed just after either the shortcuts or setting UI are fixed.

BUG=126880
TEST=manual test (even with high-DPI and external display)

Review URL: https://chromiumcodereview.appspot.com/10388141