Refactor GetCrashSignalFD to be more generic.
Enables us to pass more file descriptors down to the child process. This
will be used by the Android port.

As part of this, I cleaned up a
ChildProcessLauncher::Context::LaunchInternal to be a little more
generic. This can result in a little extra work but cleans up the
zygote/non-zygote code paths.


Review URL: https://chromiumcodereview.appspot.com/10584007