Drop frontbuffers with ui-use-gpu-process, synchronized with browser, decoupled from backbuffer drop.
On Aura, every time a tab is backgrounded, RenderWidgetHostViewAura will reset its handle to the front surface.  If that tab is foregrounded again, that front surface will not be used until synchronizing with the gpu process to make sure that surface is still available.  By doing this, the gpu process knows when it is safe to discard the front surface.

RWHVA sends a FrontSurfaceIsProtected(bool, int) message to the gpu process to keep it informed about front surface protection.  The int is a state-of-the-world identifier to protect from ABA issues.  RWHVA delays sending FrontSurfaceIsProtected(false) until after the current surface is certain to not be in use, namely after the compositor finishes the current frame and the browser thumbnailer is complete.


BUG=112842
TEST=Manual


Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=142408

Committed: http://src.chromium.org/viewvc/chrome?view=rev&revision=142541

Review URL: https://chromiumcodereview.appspot.com/10052018