This adds third party Drive app support to ChromeOS.
Drive can have third party web applications associated with documents,
and we want to support that on ChromeOS.  This adds a task generator
and task executor implementation that can handle drive apps.

It maps the application suffixes to the given app, and tells the UI
what name to display for them.

TBR=zelidrag@chromium.org
BUG=chromium:126895
TEST=Ran on device, ran new unit tests.

Review URL: https://chromiumcodereview.appspot.com/10577029