Make SendGeolocationPermissionResponse execute on UI thread.
SendGeolocationPermissionResponse can be called on the IO thread from
GeolocationDispatcherHostImpl::OnRequestPermission if there is no
geolocation_permission_context_ (as is the case with content shell).

But, SendGeolocationPermissionResponse calls
RenderViewHostImpl::FromID which must be called on the UI thread, so
we need to ensure that SendGeolocationPermissionResponse is called on
the UI thread too.

BUG=134116
TEST=Geolocation Browser Tests.


Review URL: https://chromiumcodereview.appspot.com/10644008