Update pyauto's scroll.js.
This makes it match the technique used by the Android scroll tests so that
they can be shared. It also paves the way for the real world page scrolling
benchmark by changing:

- The old code scrolled twice and discarded the first scroll. Tile caching
means the first scroll is very interesting. Now both the initial and
repeat scrolls are recorded. Repeat is still used in this patch.
- The old code used Date.now() instead of the requestAnimationFrame
timestamp. This method is more accurate.
- The old code measured only average FPS. This code measures average as
well as the number that don't complete within 60FPS.
- The old code measured FPS, but the GPU team likes to discuss frame time.

For now, the format of the results is unchanged -- FPS.

BUG=134057
TEST=chrome/test/functional/perf.py perf.ScrollTest.testTextScroll

Review URL: https://chromiumcodereview.appspot.com/10575028