Add a tool for querying tryjob statuses
This is relatively complex to implement because the Buildbot JSON API
does not provide a way to look up a trybot's status given its name
(that is, the name that "git try" or "gcl try" uses when queuing the
tryjob).  Instead, we have to linearly scan Buildbot's list of recent
tryjobs to find one with a matching name.

I'm adding this to the "parasite" directory because I intend to rework
parasite to use this code.  Parasite can currently query tryjob
statuses, but its implementation is incomplete: it will miss a tryjob
number -> name mapping if it never sees the job in the builder's
"currentBuilds" list, which is likely to happen if a job completes in
under a minute.

BUG=http://code.google.com/p/nativeclient/issues/detail?id=2862
TEST=tryjob_query_test.py

Review URL: https://chromiumcodereview.appspot.com/10639009