Allow ChromeOS file selection dialog to be shown from shell windows.
This enables the fileSystem APIs to work on ChromeOS. With this change
the dialogs are shown full screen when invoked from shell windows.

BUG=133594
TEST=Test that file selection dialog can be used successfully on
ChromeOS from a range of uses, including loading an extension and
saving a file you have downloaded. Check that the dialog looks the
same as it always has.

Review URL: https://chromiumcodereview.appspot.com/10584026