Don't use a GL extensions query to test for async swap buffers
A GL context supports swap buffers callbacks iff it is a command buffer context.
RenderViewImpl constructs a command buffer or non-command buffer context based off
of only a command line flag, so that information is sufficient.

This is mostly a drive-by cleanup, but it also improves the startup perf of
the single threaded compositor since the first extension query is synchronous
to the GPU process

BUG=none


Review URL: https://chromiumcodereview.appspot.com/9619033