Pass OAuth token refresh errors back to the caller.
This allows the app to advise the user to sign back in to Chromoting only when
we believe it will actually help. As a bonus, the unhelpfully-named
ERROR_GENERIC is now gone--apart from OAuth refresh failures, it was being used
for edge-cases where we don't expect anything to go wrong (these now use the
more suitably-named ERROR_UNEXPECTED).

As part of fixing this, I have cleaned up some OAuth call points that were
written before callWithToken existed, and which weren't using it.

BUG=122899,130794
TEST=Corrupt the oauth2-refresh-token value in HTML Local Storage

Review URL: https://chromiumcodereview.appspot.com/10579012