Remove --disable-interactive-form-validation flag and about:flags UI for it.
We introduced this flag as a workaround for web sites which are not compatible
with HTML5 form validation. However we have received just one bug report [1]
caused by this feature since January 2011.

It's safe to remove the flag because the feature didn't make many
incompatibility issues.

[1] http://webkit.org/b/44436

BUG=none
TEST=none


Review URL: https://chromiumcodereview.appspot.com/10640016