Make the metrics logic for cutting a new log and uploading an old log clearer
Separate the flow of either uploading an old log or cutting and uploading a new
log when the callback is triggered. This makes it clearer what each is doing,
and fixes a longstanding odd behavior where every time an old log is uploaded,
the histogram gathering that's meant for just before closing a log is run.

BUG=None
TEST=Metrics should continue to work.


Review URL: https://chromiumcodereview.appspot.com/10583015