Delete the Win and Mac coverage bots.
They appear to have been broken for at least 9 months. The Windows one
also need to be rebuilt. (we don't want Vista machines anymore if possible), and
it does not seem like a good use of time and resources to rebuild machines
that don't add value.

We can provide more resources if someone can commit time to fixing them and
promoting them to the rest of the team.

For now we'll need to rely on the linux bot, which is still working fine.

Review URL: https://chromiumcodereview.appspot.com/10668007