Add MetricsService::OnAppEnterBack/Foreground
On Android, Chrome may be sent to the background whenever the user decides to
look at another app or answer a phone call (e.g.).  In this state, Chrome is
effectively paused and may be resumed by the user when Chrome is reopened,
but may be killed by Android at any time before that point.

This CL introduces functions which get triggered before and after we are sent
to the background.
* When we are sent to the background, we flush all current logs to disk but
  keep recording metrics in case we are brought back.  We also stop the scheduler
  from trying to upload logs to avoid consuming bandwidth.
* If Chrome is brought back into the foreground, we again resume trying to upload
  any current logs.

BUG= b/6705461 (internal)
TEST=


Review URL: https://chromiumcodereview.appspot.com/10605004