Merge 144339 - sync: Set explicit protocol_version on commit
The use of Message::Clear() function in commit.cc was unsetting this
field.  This change explicitly sets the field to the current default
value following the Clear().  It also adds a test assertion to verify
that the field is properly set.

BUG=134769,134715
TEST=Run against the test server and verify the field is in the logs.

Review URL: https://chromiumcodereview.appspot.com/10679010

TBR=rlarocque@chromium.org
Review URL: https://chromiumcodereview.appspot.com/10687008