bluetooth: fix discovered device removal issues
The DeviceRemoved signal indicates that a device no longer has an object
path, it does not indicate that a discovered device is no longer visible
to the adapter. Some devices on receiving an inquiry during discovery
make an SDP connection to the Chromebox in return, this was causing their
removal from the discovered devices set due to the connection.

Fix by adding a new IsVisible() member to BluetoothDevice, this indicates
whether or not a device is present during discovery and is completely
orthogonal to IsPaired(). Both may be true, both may be false, or one or
the other may be set.

A BluetoothDevice structure is no longer freed on DeviceRemoved if it
is visible, instead the object path is simply removed and observers
notified of a change, rather than a removal. Likewise on
DeviceDisappeared clear the visible flag and inform observers of a change
if the device is paired.

Clean up the other observer methods at the same time so that they are
always consistent and clean up BluetoothDevice creation to remove the
bound/unbound distinction and leave that up to BluetoothAdapter.

BUG=chromium-os:31115,chromium-os:31629
TEST=made SDP connection to Chromebox from discovered desktop

Change-Id: I6485c9933ec48f66c479854c524bd5c90c917a32


Review URL: https://chromiumcodereview.appspot.com/10535080