Relanding r144496 which was reverted in r144503.
Fix several shutdown/error-path bugs in VAVDA and avoid unnecessary MakeCurrent() calls

- GpuVideoDecodeAcceleratorHost is already AddRoute'd to
GpuChannelHost on creation, so
CommandBufferProxyImpl::OnChannelError doesn't have to call
GVDAH::OnChannelError as well (was causing double-notifications
of errors to GVDAH).

- GpuVideoDecodeAcceleratorHost::Send() would DCHECK that no
error had previously been reported on the channel, but that's
incorrect because its client will likely still have async
posted tasks from before the error notifications. Now instead
of DCHECK'ing simply early-return after an error was seen.

- gpu_video_decode_accelerator.cc:MakeDecoderContextCurrent()
used to unconditionally bind the GpuCommandBufferStub* into its
closure, but nothing guarantees the lifetime of the stub; in
fact context loss will cause the stub to be deleted so this
could crash as well. Now take a WeakPtr to the stub and check
its validity before every use (and return false if invalid to
allow clients to avoid assuming context currency).

- GpuVideoDecodeAccelerator is now a
GpuCommandBufferStub::DestructionObserver to avoid attempting
further work once the stub is about to be deleted.

- Removed a bunch of unnecessary MakeCurrent() calls; VAAPI_*
calls don't need it, and are sometimes made on the decoder (not
main) thread, from which it is incorrect to call GL-related
methods such as MakeCurrent() anyway.

BUG=134020
TEST=manual playback of multiple HW-decoded .mp4's doesn't crash anything when different tabs are closed.
TBR=posciak,piman

Review URL: https://chromiumcodereview.appspot.com/10692016