bluetooth: use a shared default BluetoothAdapter
Change references to BluetoothAdapter from a scoped_ptr to a
scoped_refptr so that all users share a single instance; manage with
a WeakPtr so that we don't hold that instance around when nobody uses it.

This aids debugging since there will only be one copy of the log messages,
and makes it possible for future CLs to deal with multiple StartDiscovery
calls and implement metrics tracking.

BUG=chromium-os:31117
TEST=verified bluetooth lifecycle, and shut down chrome after

Change-Id: I754379f2ef4ac07846c99502523a618d54784df0


Review URL: https://chromiumcodereview.appspot.com/10541086