Validate the video codec profile coming into the DXVA decoder and only allow the ones we support. Currently we only support H264 baseline/main/HD. H264 HD videos are janky at times when decoded via the DXVA decoder. Will continue working on improving that.
Fixed a bug in the decode process where we were sending the MFT_MESSAGE_NOTIFY_START_OF_STREAM
message for every input packet. As per docs this should be sent at the start of an input stream,
which is once for our case.

We also set the MF_MT_INTERLACE_MODE property to MFVideoInterlace_MixedInterlaceOrProgressive as per msdn recommendations in the input media type.


BUG=none
R=fischman
Review URL: https://chromiumcodereview.appspot.com/10682010