Add alternate New Tab Page for Instant Extended work
Adds alternate New Tab Page for Instant Extended work. This is a duplicate of
the "ntp4" version of the new tab page with new_tab.html and new_tab_theme.css
forked as a starting point. This alternate page is enabled with the
--enable-instant-extended-api flag.

This is a more minimal approach than http://codereview.chromium.org/10636031/.
Original files are now referenced (not copied) to keep "sizes bots" happy.

BUG=133529
TEST=Manual
R=estade@chromium.org


Review URL: https://chromiumcodereview.appspot.com/10690020