Reduce calls to BluetoothServiceRecord constructor.
Refactor all access to service records through GetServiceRecords.  This
is slightly less efficient (as ConnectToService and ProvidesServiceWithName
can no longer early exit), but creates a consistent pattern.

If the efficiency is essential in the future, we could refactor
GetServiceRecords, ConnectToService and ProvidesServiceWithName to all
call a common function which calls a callback for each service record.

TEST=still compiles
BUG=none


Review URL: https://chromiumcodereview.appspot.com/10697014