Rewrite DownloadsApiTest in C++. This uses much less magic than the javascript version, so it's more robust and easier to debug and fix.
Aaron is the primary reviewer since most of this is extensions-specific.
Randy, none of this is particularly downloads-specific. You're welcome to review or mute.

ExtensionDownloadsEventRouter sends a chrome::NOTIFICATION_EXTENSION_DOWNLOADS_EVENT notification whenever it fires an event.
DownloadsEventsListener is-a content::NotificationObserver that listens for chrome::NOTIFICATION_EXTENSION_DOWNLOADS_EVENT and logs them. DEL facilitates waiting for specific events by running the message loops, optionally with a timeout to prevent hitting the 45s hard timeout. If a WaitFor() call times out, then it prints the events that it was waiting for along with the events that it received to aid debugging.
BrowserContext::GetFileSystemContext() is used to create HTML5 FileSystem Files.
Empty extensions are used because DownloadsDownloadFunction relies on its host permissions mechanisms.

I considered grouping all of the new TEST_Fs in a single large TEST_F in order to amortize start-up cost, but that would have required a comment explaining how to disable sub-sections, and it would have complicated time-outs. Besides, most bots can parallelize.

Review URL: https://chromiumcodereview.appspot.com/10542038