Create a PPAPI host for new resource message routing.
This infrastructure will be used in the renderer and in the browser as the backing for new resources.

The PpapiHost object doues the routing for the resource messages, and also has hooks for the embedder (the renderer or the browser) to create resources. This adds a content_renderer factory which currently does nothing (we'll add most of the resources here).

BUG=
TEST=

Review URL: https://chromiumcodereview.appspot.com/10572040