Fix a bug in file_util::Delete() where symlinks are not handled right
With this patch, file_util::Delete() uses lstat64() instead of stat64(), so that Delete() checks the existence of the symbolic link itself not a file pointed by the symlink. This allows us to delete a symbolic link when it points to a non-existent file. 

BUG=119430
TEST=base_unittests::FileUtilTest.* passes.

Review URL: https://chromiumcodereview.appspot.com/10690047