gdata: Remove invalid files in the cache directories
This is a reland of crbug.com/144993, which was reverted due to
memory error detected by valgrind. This version fixed the issue.

In theory, we shouldn't have any invalid files left in the
cache directories, but things can go wrong and invalid files
may be left if the device shuts down unexpectedly, for instance.
Besides, it's good to be defensive.

BUG=134862
TEST=added unit tests

Review URL: https://chromiumcodereview.appspot.com/10693020