Reuse video capture buffers when we have no free slots.
When the media layer sends a new buffer to Pepper, the Pepper object will save the buffer and notify the media system that the buffer can be re-used. But if there are no free buffer slots, we never tell the media layer the buffer can be reused. This causes a leak of buffers and the video will freeze.

TEST=see bug
BUT=135216

Review URL: https://chromiumcodereview.appspot.com/10689080