sync: Remove GetPossibleLastChildForTest()
Remove GetPossibleLastChildForTest().  Its only use was in
GetLastChildForTest() to make an informed guess as to which entry was
the last one under a given parent.  That function will return the same
results if we use GetPossibleFirstChild() to initialize our guess.  In
theory this new approach will be slower, but in practice the change will
be imperceptible.

BUG=135483
TEST=

Review URL: https://chromiumcodereview.appspot.com/10703064