sync: Move MarkForSyncing to mutable_entry.cc
Prior to this change, the function was declared in mutable_entry.h and defined
in syncable_util.cc.  This was the result of an error in r143218, which split
syncable.cc into several files.  This commit fixes the error by moving the
defintion of the function.

BUG=103332
TEST=


Review URL: https://chromiumcodereview.appspot.com/10696083