Extension Docs Server: Function Templates
Note: some of the templates are very ugly and hard to read. Particularly,
the type.html template. This is because it may be surrounded in a <pre></pre>,
which keeps the formatting. This is unfortunate.

These templates are used in:
https://chromiumcodereview.appspot.com/10688002/

BUG=133772
TEST=none

Review URL: https://chromiumcodereview.appspot.com/10679013
Patch from Christopher Hebert <hebert.christopherj@chromium.org>.