sync: Fix and test delete of server-unknown items
The fix we settled on for issue 125381 was to unset the unsynced bit for
items which were created and deleted locally, before the server was ever
iuformed of their existence.  The server doesn't like being notified of
deletions to items it never knew existed, as we learned with
issue 132905.

It turns out that the syncapi was deleting items in a way that bypassed
the originally attempted fix for issue 125381 in r137476.  This shows
that the testing of the original fix was inadequate.  This commit fixes
the bug and improves testing, with the hope of preventing these bugs
from reoccuring in the future.  The fix is so subtle that I doubt
anything other than lots and lots of tests will be able to prevent
regressions.

BUG=134934,125381
TEST=


Review URL: https://chromiumcodereview.appspot.com/10703060