Don't delete cookies and other non-protected data unless explicitly requested
The clear browser data dialog has now a checkbox "[ ] Clear hosted app data". This CL ensures that cookies, server-bound certificates and plugin data are not deleted by the mere fact that this checkbox is ticked. These data types are not storage protected and are therefore bound to the "[ ] Delete cookies and other site and plug-in data" checkbox.

BUG=134983
TEST=see bug

Review URL: https://chromiumcodereview.appspot.com/10689070