Add command 'merge' to isolate.py. Fix merge with symlinks.
Command 'merge' reads a trace and merges it back into the original .isolate
file.  This is a shortcut to simplify the workflow when updating a .isolate file
automatically.

When using merge_isolate.py, if a symlinked file was listed as an untracked
dependency, it would be incorrectly upgraded to a tracked dependency.

TBR=cmp@chromium.org
NOTRY=true
BUG=
TEST=


Review URL: https://chromiumcodereview.appspot.com/10703098