Refactor CreateOrUpdateChromeShortcuts() for a better public API.
This is necessary as the follow-up patches will need a clean way to call CreateOrUpdateChromeShortcuts() while knowing the outcome which was kind of vague before (and would have required passing in a fake InstallStatus to simulate the desired behaviour...).

BUG=132825, 127544, 131415
TEST=Install Chrome at user and system-level with and without create_all_shortcuts and verify that shortcut creation is as expected.
Over-install this custom installer over chrome-dev after having deleted its shortcuts and verify they are not recreated.


Review URL: https://chromiumcodereview.appspot.com/10692034