Make the directx sdk and the regular sdk path a gyp variable
For the directx it probes in src\third_party\directxsdk if you have the dxsdk there and
for the regular sdk it probems in src\third_party\platformsdk_win8

If it does not find them it takes an educated guess.

The gyp variables windows_sdk_path and directx_sdk_path can be overriden from the environment if
the defaults are wrong.

BUG=none
TEST=chrome builds
Review URL: https://chromiumcodereview.appspot.com/10690072