Make EventBindings use static routed methods.
This change makes EventBindings use NativeHandler::RouteStaticFunction()
instead of NativeHandler::RouteFunction() because use of the latter seems to
cause use after free errors (see bug).

This change was originally done in r132850 but it was reverted in r133253
because it introduced many crashes.

In r145484 as part of a bigger change (filtered events) the events were
made to use dynamic routing again and many crashes ensued again.


BUG=136232
TBR=kalman@chromium.org

Review URL: https://chromiumcodereview.appspot.com/10694108